+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+# 1.7.15 (2022-03-24)
+
+## Project:
+
+- Improve support for sequence names when using the project settings
+   In the Project Settings panel the shot name template has been replaced by 3 identifiers,
+   one for the project or act, one for the sequence and one for the shot.
+
+   In the main panel the name of the sequence, when the project settings are used, is now
+   set thanks to a dedicated dialog box.
+
+## Tools:
+
+- Improve the shot tool named Create Specified Number of Shots
+
+## Fixed:
+
+- FPS variable issue when exporting edit file
+
+
 # 1.7.10 (2022-03-03)
 
 ## Fixed: